Skill ut dokumentÅrsakTyper og barnIBrevÅrsak fra Dokumentutsending
Flytter DokumentÅrsak-enum og dokumentÅrsak-tekster ut av
DokumentutsendingContext til egen fil, og BarnIBrevÅrsak/
barnIBrevÅrsakTilTittel/finnBarnIBrevÅrsak ut av DokumentutsendingSkjema.
Ren uttrekk uten funksjonell endring, slik at kommende React Hook Form-
komponenter kan importere disse uten å avhenge av contexten.
