+++ src/frontend/sider/Fagsak/Dokumentutsending/barnIBrevÅrsak.ts
@@ -1,0 +1,25 @@
+import { DokumentÅrsak } from './dokumentÅrsakTyper';
+
+export enum BarnIBrevÅrsak {
+    BARN_SØKT_FOR,
+    BARN_BOSATT_MED_SØKER,
+}
+
+export const barnIBrevÅrsakTilTittel: Record<BarnIBrevÅrsak, string> = {
+    [BarnIBrevÅrsak.BARN_SØKT_FOR]: 'Hvilke barn er søkt for?',
+    [BarnIBrevÅrsak.BARN_BOSATT_MED_SØKER]: 'Hvilke barn er bosatt med søker?',
+};
+
+export const finnBarnIBrevÅrsak = (årsak: DokumentÅrsak | undefined): BarnIBrevÅrsak | undefined => {
+    switch (årsak) {
+        case DokumentÅrsak.KAN_SØKE_EØS:
+        case DokumentÅrsak.TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HAR_FÅTT_EN_SØKNAD_FRA_ANNEN_FORELDER:
+        case DokumentÅrsak.TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_VARSEL_OM_REVURDERING:
+        case DokumentÅrsak.TIL_FORELDER_OMFATTET_NORSK_LOVGIVNING_HENTER_IKKE_REGISTEROPPLYSNINGER:
+            return BarnIBrevÅrsak.BARN_SØKT_FOR;
+        case DokumentÅrsak.KAN_HA_RETT_TIL_PENGESTØTTE_FRA_NAV:
+            return BarnIBrevÅrsak.BARN_BOSATT_MED_SØKER;
+        default:
+            return undefined;
+    }
+};
